Refactor api-gateway
