+++ .gitattributes
@@ -1,2 +1,2 @@
-pkg/cmd/api/spec/api.swagger.json linguist-generated=true
-pkg/cmd/api/spec/static.go linguist-generated=true
+pkg/apigateway/spec/api.swagger.json linguist-generated=true
+pkg/apigateway/spec/static.go linguist-generated=true
