+++ pkg/apigateway/spec/makestatic.go
