+++ pkg/apigateway/spec/static.go
