+++ pkg/apigateway/swagger-ui/makestatic.go
