+++ pkg/apigateway/swagger-ui/static.go
