+++ pkg/apigateway/swagger-ui/swagger-ui-standalone-preset.js
