+++ pkg/apigateway/swagger-ui/version.txt
